requirements.txt
transformers
torch
datasets
pillow
numpy
pdf2image
pytesseract
langchain-groq
streamlit
python-dotenv